senior_project/requirements.txt
asgiref==3.5.0
Django==4.0.3
django-address==0.2.6
djangorestframework==3.13.1
psycopg2==2.9.3
python-decouple==3.6
python-stdnum==1.17
pytz==2022.1
sqlparse==0.4.2
